Moved 2d validation to correct file
